requirements.txt
# MANDATORY
python-dotenv
#
flask
flask_mail
flask_sqlalchemy
pysqlite3
